requirements.txt
customtkinter
pypdf
pymupdf
pdf2docx
Pillow
reportlab
openpyxl
pandas
docx2pdf
pdfplumber
comtypes
pillow-heif
extract-msg
python-pptx
xhtml2pdf
deep-translator
ebooklib
beautifulsoup4
requests
rembg
opencv-python
pytesseract
psd-tools
pillow-avif-plugin
moviepy
svglib
qrcode[pil]
pyzbar
